Update snapshot, move styles to component files
