+++ packages/grafana-ui/src/components/Tabs/TabsBar.story.tsx
@@ -18,9 +18,9 @@ export default {
 };
 
 const tabs = [
-  { label: '1st child', key: 'first', hide: false, active: true },
-  { label: '2nd child', key: 'second', hide: false, active: false },
-  { label: '3rd child', key: 'third', hide: false, active: false },
+  { label: '1st child', key: 'first', active: true },
+  { label: '2nd child', key: 'second', active: false },
+  { label: '3rd child', key: 'third', active: false },
 ];
 
 export const Simple = () => {
